Amini GeoFM Decoding the Field Challenge/requirements.txt
affine==2.4.0
attrs==25.3.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
cloudpickle==3.1.1
colorama==0.4.6
dask==2025.5.1
fsspec==2025.5.1
idna==3.10
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
locket==1.0.0
numpy==2.3.1
odc-geo==0.4.10
odc-loader==0.5.1
odc-stac==0.4.0
packaging==25.0
pandas==2.3.0
partd==1.4.2
pyparsing==3.2.3
pyproj==3.7.1
pystac==1.13.0
pystac-client==0.8.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
shapely==2.1.1
six==1.17.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
xarray==2025.6.1
